InvocationId is now a ushort. Addressed issue with InvocationIds wrapping around and re-using an active id.

+++ src/NexNet/Messages/InvocationResultMessage.cs
@@ -28,7 +28,7 @@ internal partial class InvocationResultMessage : IMessageBase
     }
 
     [MemoryPackOrder(0)]
-    public int InvocationId { get; set; }
+    public ushort InvocationId { get; set; }
 
     [MemoryPackOrder(1)]
     public StateType State { get; set; }
